Module Asetmap.Map

Maps.

Maps

module type S = sig ... end

Signature for maps with keys of a totally ordered type.

module type S_with_key_set = sig ... end

Signature for maps with keys of a totally ordered type equiped with a set.

module Make (Ord : Map.OrderedType) : S with type key = Ord.t and type 'a t = 'a Map.Make(Ord).t

Make (Ord) is a map data structure with keys of the totally ordered type Ord.

module Make_with_key_set (Ord : Map.OrderedType) (Key_set : Set.S with type elt = Ord.t) : S_with_key_set with type key = Ord.t and type key_set = Key_set.t and type 'a t = 'a Map.Make(Ord).t

Make_with_key_set (Ord) (Key_set) is a map data structure with keys of the totally ordered type Ord and key sets Key_set.