Module Asetmap.Set

Sets.

Sets

module type S = sig ... end

Signature for sets of values of a totally ordered type.

module Make : functor (Ord : Stdlib.Set.OrderedType) -> S with type elt = Ord.t and type t = Stdlib.Set.Make(Ord).t

Make (Ord) is a set data structure for values of the totally ordered type Ord.